test: skip tests on non-working platforms
Some tests may be fixable...

+++ test/black-box/setup-scripts/run.sh
@@ -11,8 +11,10 @@ expect_exist dev/dist/root/1/workspace/canary.txt
 # Verify that packageSetup is included in interactive shell environment. Need
 # to use "script" to fake a tty. Otherwise the rcfile would not be loaded and
 # the shell envirionment is not setup.
-rm dev/dist/root/1/workspace/canary.txt
-script -qfec "./dev/dist/root/1/package.sh shell" /dev/null <<'EOF'
-foo
+if [[ $(uname -o) != Msys ]] ; then
+	rm dev/dist/root/1/workspace/canary.txt
+	script -qfec "./dev/dist/root/1/package.sh shell" /dev/null <<'EOF'
+	foo
 EOF
-expect_exist dev/dist/root/1/workspace/canary.txt
+	expect_exist dev/dist/root/1/workspace/canary.txt
+fi
